Modify CSS to simplify album-metadata grid layout
The changes made in the CSS reduce the grid columns in the album-metadata from four to three. These modifications help to remove redundant div elements and thus, enhance the structure of the layout. This also possibly improves the loading speed and the ease of maintenance without affecting the visual output negatively.